chore: message.nim - set max message size to 150KiB according to spec (#2298)
* message.nim: set max message size to 150KiB according to spec

Using KiB instead of KB because that seems more aligned with
the actual default defined in nim-libp2p (1024 * 1024)

Spec details: https://rfc.vac.dev/spec/64/#message-size

* test_protocol.nim: align test to current WakuMessage limit
* test_waku_client.nim: adapt test to MaxWakuMessageSize change
* make maxMessageSize configurable for wakunode2
* wakunode2 app now accepts max-num-bytes-msg-size with KiB, KB, or B units
* testlib/wakunode.nim: set maxMessageSize: "1024 KiB"
* test_waku_client.nim: remove duplicate check in "Valid Payload Sizes"
* set DefaultMaxWakuMessageSizeStr as the only source of truth
* external_config.nim: rename max-num-bytes-msg-size -> max-msg-size

+++ tests/resources/payloads.nim
@@ -70,7 +70,7 @@ proc getSampleJsonList*(): JsonNode =
   ]
 
 
-proc getByteSequence*(bytesNumber: int): seq[byte] =
+proc getByteSequence*(bytesNumber: uint64): seq[byte] =
   result = newSeq[byte](bytesNumber)
   for i in 0 ..< bytesNumber:
     result[i] = cast[byte](i mod 256)
